Separate test files, update with read_config
